requirements.txt
langchain-openai~=0.3.32
langchain~=0.3.25
langchain-community~=0.3.24
langchain-text-splitters~=0.3.8
langchain-chroma==0.2.5
pandas~=2.3.1
numpy~=2.2.6
scikit-learn~=1.6.1
python-dotenv~=1.1.0
faiss-cpu~=1.12.0
pypdf==5.5.0
streamlit

